+++ feincms3_sites/middleware.py
@@ -1,4 +1,5 @@
 import contextvars
+import warnings
 from contextlib import contextmanager
 
 from django.conf import settings
@@ -8,7 +9,6 @@ from django.utils import translation
 from django.utils.cache import patch_vary_headers
 
 from feincms3.apps import AppsMixin, apps_urlconf
-from feincms3.utils import concrete_model
 
 
 _current_site = contextvars.ContextVar("current_site")
@@ -26,6 +26,14 @@ def current_site():
 
 
 def apps_urlconf_for_site(site=None):
+    warnings.warn(
+        "apps_urlconf_for_site is deprecated since feincms3.apps.apps_urlconf"
+        " now automatically filters by site.",
+        Warning,
+        stacklevel=2,
+    )
+    from feincms3.utils import concrete_model
+
     page_model = concrete_model(AppsMixin)
     fields = ("path", "application", "app_instance_namespace", "language_code")
     apps = (
@@ -52,6 +60,14 @@ def site_middleware(get_response):
 
 def apps_middleware(get_response):
     from .models import Site
+
+    warnings.warn(
+        "feincms3_sites.middleware.apps_middleware is deprecated."
+        " Simply use feincms3_sites.middleware.site_middleware and"
+        " feincms3.apps.apps_middleware directly after each other.",
+        Warning,
+        stacklevel=2,
+    )
 
     def middleware(request):
         request.site = Site.objects.for_host(request.get_host())