+++ CHANGELOG.rst
@@ -5,6 +5,11 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+- Deprecated ``apps_urlconf_for_site`` and ``apps_middleware`` since
+  feincms3's ``apps_middleware`` now automatically does the right thing
+  when used after ``site_middleware``.
+
+
 `0.7`_ (2019-02-06)
 ~~~~~~~~~~~~~~~~~~~